# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "WhatsApp Web Panel API",
    "version": "1.0.0",
    "description": "Cookie tabanli oturum (express-session) + CSRF korumali API. /api altindaki mutating (POST/PUT/DELETE) endpoint'ler icin CSRF header gonderilmelidir."
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "whatsapp.sid"
      },
      "csrfToken": {
        "type": "apiKey",
        "in": "header",
        "name": "X-CSRF-Token"
      },
      "xsrfToken": {
        "type": "apiKey",
        "in": "header",
        "name": "X-XSRF-TOKEN"
      }
    },
    "parameters": {
      "AccountIdHeader": {
        "name": "X-Account-Id",
        "in": "header",
        "required": false,
        "schema": { "type": "string" },
        "description": "Multi-account icin aktif hesabi secmekte kullanilir."
      }
    },
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "required": ["error"],
        "properties": {
          "error": { "type": "string" },
          "requestId": { "type": "string" }
        }
      },
      "SuccessResponse": {
        "type": "object",
        "required": ["success"],
        "properties": {
          "success": { "type": "boolean" }
        }
      },
...
    }
  },
  "paths": {
...
    "/api/status": {
      "get": {
        "tags": ["WhatsApp"],
        "summary": "WhatsApp durum + message stats",
        "security": [{ "sessionCookie": [] }],
        "parameters": [{ "$ref": "#/components/parameters/AccountIdHeader" }],
        "responses": {
          "200": { "description": "Status", "content": { "application/json": { "schema": { "type": "object" } } } },
          "401": { "description": "Unauthorized", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }
        }
      }
    },
...
    "/api/chats/{chatId}/messages": {
      "get": {
        "tags": ["Chats"],
        "summary": "Chat messages + tags + notes",
        "security": [{ "sessionCookie": [] }],
        "parameters": [
          { "$ref": "#/components/parameters/AccountIdHeader" },
          { "name": "chatId", "in": "path", "required": true, "schema": { "type": "string" } },
          { "name": "limit", "in": "query", "required": false, "schema": { "type": "integer" } },
          { "name": "offset", "in": "query", "required": false, "schema": { "type": "integer" } }
        ],
        "responses": { "200": { "description": "Messages", "content": { "application/json": { "schema": { "type": "object" } } } } }
      }
    },
    "/api/messages": {
      "get": {
        "tags": ["Messages"],
        "summary": "Messages list (paginated)",
        "security": [{ "sessionCookie": [] }],
        "parameters": [
          { "$ref": "#/components/parameters/AccountIdHeader" },
          { "name": "limit", "in": "query", "required": false, "schema": { "type": "integer" } },
          { "name": "offset", "in": "query", "required": false, "schema": { "type": "integer" } }
        ],
        "responses": { "200": { "description": "Messages", "content": { "application/json": { "schema": { "type": "object" } } } } }
      }
    },
    "/api/messages/search": {
      "get": {
        "tags": ["Messages"],
        "summary": "Messages search",
        "security": [{ "sessionCookie": [] }],
        "parameters": [
          { "$ref": "#/components/parameters/AccountIdHeader" },
          { "name": "q", "in": "query", "required": true, "schema": { "type": "string" } }
        ],
        "responses": { "200": { "description": "Results", "content": { "application/json": { "schema": { "type": "array", "items": { "type": "object" } } } } } }
      }
    },
...
    "/api/contacts/{chatId}/tags": {
      "get": {
        "tags": ["Contacts"],
        "summary": "Contact tags",
        "security": [{ "sessionCookie": [] }],
        "parameters": [
          { "$ref": "#/components/parameters/AccountIdHeader" },
          { "name": "chatId", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Tags", "content": { "application/json": { "schema": { "type": "array", "items": { "type": "object" } } } } }
        }
      },
      "post": {
        "tags": ["Contacts"],
        "summary": "Contact tag ekle",
        "security": [{ "sessionCookie": [], "csrfToken": [] }],
        "parameters": [
          { "$ref": "#/components/parameters/AccountIdHeader" },
          { "name": "chatId", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["tag_id"],
                "properties": { "tag_id": { "type": "integer" } }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Success", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/SuccessResponse" } } } }
        }
      }
    },
    "/api/contacts/{chatId}/tags/{tagId}": {
      "delete": {
        "tags": ["Contacts"],
        "summary": "Contact tag kaldir",
        "security": [{ "sessionCookie": [], "csrfToken": [] }],
        "parameters": [
          { "$ref": "#/components/parameters/AccountIdHeader" },
          { "name": "chatId", "in": "path", "required": true, "schema": { "type": "string" } },
          { "name": "tagId", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "responses": {
          "200": { "description": "Success", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/SuccessResponse" } } } }
        }
      }
    },
...
    "/api/reports/overview": {
      "get": {
        "tags": ["Reports"],
        "summary": "Reports overview",
        "security": [{ "sessionCookie": [] }],
        "parameters": [
          { "$ref": "#/components/parameters/AccountIdHeader" },
          { "name": "start", "in": "query", "required": false, "schema": { "type": "integer", "format": "int64" } },
          { "name": "end", "in": "query", "required": false, "schema": { "type": "integer", "format": "int64" } },
          { "name": "limit", "in": "query", "required": false, "schema": { "type": "integer" } }
        ],
        "responses": { "200": { "description": "Overview", "content": { "application/json": { "schema": { "type": "object" } } } }, "400": { "description": "Invalid range", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } } }
      }
    },
    "/api/reports/trends": {
      "get": {
        "tags": ["Reports"],
        "summary": "Reports trends",
        "security": [{ "sessionCookie": [] }],
        "parameters": [
          { "$ref": "#/components/parameters/AccountIdHeader" },
          { "name": "start", "in": "query", "required": false, "schema": { "type": "integer", "format": "int64" } },
          { "name": "end", "in": "query", "required": false, "schema": { "type": "integer", "format": "int64" } },
          { "name": "interval", "in": "query", "required": false, "schema": { "type": "string", "enum": ["daily", "weekly"] } }
        ],
        "responses": { "200": { "description": "Trends", "content": { "application/json": { "schema": { "type": "object" } } } }, "400": { "description": "Invalid range", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } } }
      }
    },
    "/api/reports/response-time": {
      "get": {
        "tags": ["Reports"],
        "summary": "Reports response time",
        "security": [{ "sessionCookie": [] }],
        "parameters": [
          { "$ref": "#/components/parameters/AccountIdHeader" },
          { "name": "start", "in": "query", "required": false, "schema": { "type": "integer", "format": "int64" } },
          { "name": "end", "in": "query", "required": false, "schema": { "type": "integer", "format": "int64" } },
          { "name": "limit", "in": "query", "required": false, "schema": { "type": "integer" } },
          { "name": "interval", "in": "query", "required": false, "schema": { "type": "string", "enum": ["daily", "weekly"] } }
        ],
        "responses": { "200": { "description": "Response time report", "content": { "application/json": { "schema": { "type": "object" } } } }, "400": { "description": "Invalid range", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } } }
      }
    },
    "/api/roles": {
      "get": {
        "tags": ["Roles"],
        "summary": "Roles list (admin)",
        "security": [{ "sessionCookie": [] }],
        "parameters": [{ "$ref": "#/components/parameters/AccountIdHeader" }],
        "responses": { "200": { "description": "Roles", "content": { "application/json": { "schema": { "type": "array", "items": { "type": "object" } } } } } }
      },
      "post": {
        "tags": ["Roles"],
        "summary": "Role create (admin)",
        "security": [{ "sessionCookie": [], "csrfToken": [] }],
        "parameters": [{ "$ref": "#/components/parameters/AccountIdHeader" }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object", "required": ["name"], "properties": { "name": { "type": "string" }, "description": { "type": "string" } } } } } },
        "responses": { "200": { "description": "Created", "content": { "application/json": { "schema": { "type": "object" } } } }, "409": { "description": "Already exists", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } } }
      }
    },
    "/api/roles/{id}": {
      "delete": {
        "tags": ["Roles"],
        "summary": "Role delete (admin)",
        "security": [{ "sessionCookie": [], "csrfToken": [] }],
        "parameters": [
          { "$ref": "#/components/parameters/AccountIdHeader" },
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "responses": { "200": { "description": "Success", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/SuccessResponse" } } } }, "400": { "description": "Assigned", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } } }
      }
    },
...
    "/api/users/{id}/role": {
      "put": {
        "tags": ["Users"],
        "summary": "User role update (admin)",
        "security": [{ "sessionCookie": [], "csrfToken": [] }],
        "parameters": [
          { "$ref": "#/components/parameters/AccountIdHeader" },
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object", "required": ["roleId"], "properties": { "roleId": { "type": "integer" } } } } } },
        "responses": { "200": { "description": "Success", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/SuccessResponse" } } } } }
      }
    },
...
    "/api/drive/status": {
      "get": {
        "tags": ["Drive"],
        "summary": "Drive status",
        "security": [{ "sessionCookie": [] }],
        "parameters": [{ "$ref": "#/components/parameters/AccountIdHeader" }],
        "responses": { "200": { "description": "Status", "content": { "application/json": { "schema": { "type": "object" } } } } }
      }
    },
    "/api/drive/migrate": {
      "post": {
        "tags": ["Drive"],
        "summary": "Drive migrate existing media",
        "security": [{ "sessionCookie": [], "csrfToken": [] }],
        "parameters": [{ "$ref": "#/components/parameters/AccountIdHeader" }],
        "responses": { "200": { "description": "Result", "content": { "application/json": { "schema": { "type": "object" } } } } }
      }
    },
    "/api/drive/upload": {
      "post": {
        "tags": ["Drive"],
        "summary": "Drive upload",
        "security": [{ "sessionCookie": [], "csrfToken": [] }],
        "parameters": [{ "$ref": "#/components/parameters/AccountIdHeader" }],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": ["file"],
                "properties": {
                  "file": { "type": "string", "format": "binary" }
                }
              }
            }
          }
        },
        "responses": { "200": { "description": "Uploaded", "content": { "application/json": { "schema": { "type": "object" } } } } }
      }
    },
...
  }
}